+++ src/domain/carrier/application/use-cases/create-order.ts
@@ -1,7 +1,7 @@
-import { Order } from '@/domain/carrier/enterprise/entities/order'
-import { UniqueEntityID } from '@/core/entities/unique-entity-id'
-import { InMemoryOrdersRepository } from '@test/repositories/in-memory/in-memory-orders-repository'
 import { Either, right } from '@/core/either'
+import { UniqueEntityID } from '@/core/entities/unique-entity-id'
+import { Order } from '@/domain/carrier/enterprise/entities/order'
+import { InMemoryOrdersRepository } from '@test/repositories/in-memory/in-memory-orders-repository'
 
 interface CreateOrderUseCaseRequest {
   recipientId: string
