+++ src/domain/carrier/application/use-cases/delete-recipient.spec.ts
@@ -1,7 +1,7 @@
-import { makeRecipient } from '@test/factories/make-recipient'
-import { DeleteRecipientUseCase } from './delete-recipient'
-import { InMemoryRecipientsRepository } from 'test/repositories/in-memory/in-memory-recipients-repository'
 import { UniqueEntityID } from '@/core/entities/unique-entity-id'
+import { makeRecipient } from '@test/factories/make-recipient'
+import { InMemoryRecipientsRepository } from '@test/repositories/in-memory/in-memory-recipients-repository'
+import { DeleteRecipientUseCase } from './delete-recipient'
 import { ResourceNotFoundError } from './errors/resource-not-found-error'
 
 let inMemoryRecipientsRepository: InMemoryRecipientsRepository
