+++ src/domain/carrier/application/use-cases/edit-courier.spec.ts
@@ -1,7 +1,7 @@
 import { UniqueEntityID } from '@/core/entities/unique-entity-id'
-import { EditCourierUseCase } from './edit-courier'
 import { makeCourier } from '@test/factories/make-courier'
-import { InMemoryCouriersRepository } from 'test/repositories/in-memory/in-memory-couriers-repository'
+import { InMemoryCouriersRepository } from '@test/repositories/in-memory/in-memory-couriers-repository'
+import { EditCourierUseCase } from './edit-courier'
 import { ResourceNotFoundError } from './errors/resource-not-found-error'
 
 let inMemoryCouriersRepository: InMemoryCouriersRepository
