+++ src/domain/carrier/application/use-cases/fetch-couriers.spec.ts
@@ -1,4 +1,4 @@
-import { makeCourier } from 'test/factories/make-courier'
+import { makeCourier } from '@test/factories/make-courier'
 import { InMemoryCouriersRepository } from '@test/repositories/in-memory/in-memory-couriers-repository'
 import { FetchCouriersUseCase } from './fetch-couriers'
