+++ src/domain/carrier/application/use-cases/update-order-status.spec.ts
@@ -1,9 +1,9 @@
 import { UniqueEntityID } from '@/core/entities/unique-entity-id'
-import { UpdateOrderStatusUseCase } from './update-order-status'
 import { Status } from '@/domain/carrier/enterprise/entities/order'
 import { makeOrder } from '@test/factories/make-order'
-import { InMemoryOrdersRepository } from 'test/repositories/in-memory/in-memory-orders-repository'
+import { InMemoryOrdersRepository } from '@test/repositories/in-memory/in-memory-orders-repository'
 import { ResourceNotFoundError } from './errors/resource-not-found-error'
+import { UpdateOrderStatusUseCase } from './update-order-status'
 
 let inMemoryOrdersRepository: InMemoryOrdersRepository
 let sut: UpdateOrderStatusUseCase
