+++ src/domain/notification/application/use-cases/read-notification.spec.ts
@@ -1,26 +1,28 @@
-import { InMemoryNotificationsRepository } from '@test/repositories/in-memory/in-memory-notifications-repository'
-import { ReadNotificationUseCase } from './read-notification'
-import { makeNotification } from '@test/factories/make-notification'
 import { UniqueEntityID } from '@/core/entities/unique-entity-id'
 import { ResourceNotFoundError } from '@/domain/carrier/application/use-cases/errors/resource-not-found-error'
+import { makeNotification } from '@test/factories/make-notification'
+import { InMemoryNotificationsRepository } from '@test/repositories/in-memory/in-memory-notifications-repository'
+import { ReadNotificationUseCase } from './read-notification'
 
 let inMemoryNotificationsRepository: InMemoryNotificationsRepository
 let sut: ReadNotificationUseCase
 
 describe('Read Notification [USE CASE]', () => {
   beforeEach(() => {
-    inMemoryNotificationsRepository =
-      new InMemoryNotificationsRepository()
+    inMemoryNotificationsRepository = new InMemoryNotificationsRepository()
     sut = new ReadNotificationUseCase(inMemoryNotificationsRepository)
   })
 
   it('should be able to read a notification', async () => {
-    const notification = makeNotification({}, new UniqueEntityID('notification-1'))
+    const notification = makeNotification(
+      {},
+      new UniqueEntityID('notification-1'),
+    )
 
     await inMemoryNotificationsRepository.create(notification)
 
     const result = await sut.execute({
-      notificationId: 'notification-1'
+      notificationId: 'notification-1',
     })
 
     expect(result.isRight()).toEqual(true)
@@ -28,7 +30,7 @@ describe('Read Notification [USE CASE]', () => {
 
   it('should not be able to read an inexistent notification', async () => {
     const result = await sut.execute({
-      notificationId: 'notification-1'
+      notificationId: 'notification-1',
     })
 
     expect(result.isLeft()).toEqual(true)
