+++ src/domain/notification/application/use-cases/read-notification.ts
@@ -16,7 +16,8 @@ export class ReadNotificationUseCase {
   async execute({
     notificationId,
   }: ReadNotificationUseCaseRequest): Promise<ReadNotificationUseCaseResponse> {
-    const notification = await this.notificationsRepository.findById(notificationId)
+    const notification =
+      await this.notificationsRepository.findById(notificationId)
 
     if (!notification) {
       return left(new ResourceNotFoundError())
