+++ test/factories/make-courier.ts
@@ -1,10 +1,9 @@
-import { faker } from '@faker-js/faker'
-
+import { UniqueEntityID } from '@/core/entities/unique-entity-id'
 import {
   Courier,
   CourierProps,
 } from '@/domain/carrier/enterprise/entities/courier'
-import { UniqueEntityID } from '@/core/entities/unique-entity-id'
+import { faker } from '@faker-js/faker'
 
 export function makeCourier(
   override: Partial<CourierProps> = {},
